requirements.txt
python-dotenv
kafka-python
tweepy
plotly
dash
numpy
pandas
seaborn
matplotlib
transformers
tensorflow
Flask
faust
pymongo